% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{doANOVA}
\alias{doANOVA}
\title{make n-way ANOVA and Tukey Post-hocs}
\usage{
doANOVA(frml, ds, which_levels = 1, quiet = F, post_hoc = TRUE, ...)
}
\arguments{
\item{frml}{formula. Formula to be passed to aov}

\item{ds}{data.frame. Data for aov}

\item{quiet}{logical. should the results of the anova be printed? default is FALSE}

\item{...}{further arguments to be passed to aov}

\item{which_level}{character or numeric vector. which levels to compare in
post-hoc. default = 1}

\item{just_aov}{logical. Should the post-hoc test be done? Default is TRUE}
}
\value{
ANOVA p-value and Tukey's Post-hoc p-val per comparison
}
\description{
make n-way ANOVA and Tukey Post-hocs
}
\examples{
doAnova(formula = x ~ grp * gr2, which = 'grp:grp2')
}
\author{
Henrique Cabral
}

