% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotSettings.R
\name{fillGradientPlot}
\alias{fillGradientPlot}
\title{this function sets the default colours for fill gradient plots}
\usage{
fillGradientPlot(return_cols = F, ...)
}
\arguments{
\item{return_cols}{logical. if TRUE, return only colours, not scale_fill_gradient2 object. Default is FALSE}

\item{...}{other parameters passed on to scale_fill_gradient2}
}
\value{
scale_fill_gradient2 object with standard colours (red-green)
}
\description{
this function sets the default colours for fill gradient plots
}
\examples{
 fillGradientPlot()
 fillGradientPlot(name='TITLE')
}
\author{
Henrique Cabral
}

