% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textUtils.R
\name{formatColumnNames}
\alias{formatColumnNames}
\title{Function to format column names}
\usage{
formatColumnNames(x, cap_all = TRUE)
}
\arguments{
\item{x}{string. vector of names}

\item{cap_all}{logical. Capitalize all words or only first one? Default is TRUE}
}
\value{
x with replaced '_' and capitalized first letter(s)
}
\description{
Function to format column names
}
\examples{
formatColumnNames('my_column_name')
}
\author{
Henrique Cabral
}

