% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textUtils.R
\name{formatText}
\alias{formatText}
\title{Function to format text for pretty display}
\usage{
formatText(x, cap = NULL, capitalize = "first", sep_words = "_")
}
\arguments{
\item{x}{string or vector of strings.}

\item{cap}{integer. max number of characters. Default = 20}

\item{capitalize}{character. one of 'all','none','first' (default)}

\item{sep_words}{character. separate words. Default = '_' (use NULL not to separate)}
}
\value{
string or vector of strings.
}
\description{
Function to format text for pretty display
}
\examples{
formatText('hello', capitalize = 'none')
}
\author{
Henrique Cabral
}

