% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotSettings.R
\name{ggplotSetTheme}
\alias{ggplotSetTheme}
\title{This function sets the default settings for ggplots}
\usage{
ggplotSetTheme(boost = 0, ...)
}
\arguments{
\item{boost}{integer. value by which to increase or decrease font size. Default = 0}

\item{...}{other params to be passed to theme()}
}
\value{
settings to be added to a ggplot plot
}
\description{
This function sets the default settings for ggplots
}
\examples{
 ggplotSetTheme()
 ggplotSetTheme(axis.text.x = element_text(angle = 45, hjust = 1))
}
\author{
Henrique Cabral
}

