% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textUtils.R
\name{keepAlphaNum}
\alias{keepAlphaNum}
\title{Function to keep only alphanum characters}
\usage{
keepAlphaNum(x, replace = "")
}
\arguments{
\item{x}{string.}

\item{replace}{string. replace non alpha-num characters with it. default = ''}
}
\value{
character vector
}
\description{
Function to keep only alphanum characters
}
\examples{
keepAlphaNum('hello.baby')
}
\author{
Henrique Cabral
}

