% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textUtils.R
\name{myLinker}
\alias{myLinker}
\title{Function to make link in pdf file}
\usage{
myLinker(x, x_link = NULL, cap = NULL, target = F, prefix = NULL,
  type = "markdown", keep_alphanum = TRUE)
}
\arguments{
\item{x}{Name to be displayed in link}

\item{x_link}{if missing, x is used}

\item{cap}{numeric. truncate x to CAP characters}

\item{target}{logical. is link FALSE. or link target (TRUE, default)}

\item{prefix}{character. add prefix to make link unique}

\item{type}{characte. markdown/excel (XLConnect format)}

\item{keep_alphanum}{logical. Remove non-alphanumeric characters(default: TRUE)}
}
\value{
character vector
}
\description{
Function to make link in pdf file
}
\examples{
myLinker('hellothere', x_link = 'hello_link', cap = 5)
}
\author{
Henrique Cabral
}

