% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textUtils.R
\name{num2word}
\alias{num2word}
\title{Function to convert number to words}
\usage{
num2word(x)
}
\arguments{
\item{x}{numeric vector or character vector with digits.}
}
\value{
character vector
}
\description{
Function to convert number to words
}
\examples{
num2word('245885')
}
\author{
Henrique Cabral
}

