% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textUtils.R
\name{rmNonUTF8}
\alias{rmNonUTF8}
\title{Function to remove non-UTF8 characters}
\usage{
rmNonUTF8(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
character vector
}
\description{
Function to remove non-UTF8 characters
}
\examples{
rmNonUTF8('hell\\023o')
}
\author{
Henrique Cabral
}

