% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timeIt}
\alias{timeIt}
\title{Function to time operation and print start and finish message}
\usage{
timeIt(expr, msg)
}
\arguments{
\item{expr}{an object to be evaluated and timed.}

\item{msg}{string. A message to be printed}
}
\value{
prints out message and time of process. Returns the result of expr
}
\description{
Function to time operation and print start and finish message
}
\examples{
timeIt({sum(rnorm(1000))}, 'Processing...')
}
\author{
Henrique Cabral
}

