% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuery.R
\name{getQuery}
\alias{getQuery}
\title{executes a query to an oracle database}
\usage{
getQuery(query_string, ora_file = NULL, access_file = NULL,
  network_alias = "EDWPPROD", sample_table = 3, quiet = FALSE, ...)
}
\arguments{
\item{query_string}{character. can be either a sql query, the path to an sql file or the name of a table (the latter one detected by the presence of a single word in the string)}

\item{ora_file}{character. Location to the tnsnames.ora file. If NULL (default), script will look for it in
"C:\\OracleClient\\tnsnames.ora" or in "~/.tnsnames.ora"}

\item{access_file}{list or character. Either a list with fields username and password, either a character string indicating the location of the file. If NULL (default) the script will look for the file .access_db.yaml in the home directory}

\item{network_alias}{character. If tnsnames.ora file has more than one connection, which network_alias should it use? Default is EDWPDEV}

\item{sample_table}{integer. Whether to query a sample of a table. Default is 3}

\item{quiet}{logical. Be silent? Defauls is FALSE}
}
\value{
the output of the query as a data.table object
}
\description{
executes a query to an oracle database
}
\details{
You can also insert data into a table by passing an insert query and a data structure to your input files (see example below)
}
\examples{
 getQuery("select * from MYTABLE")
 getQuery("sql/my_query.sql")
 getQuery("insert into TABLE_NAME values(:1, :2)", data = ds)
}
\author{
Henrique Cabral
}

